Main file
This file uses a combination of grep statements (sometimes piped to each other) in order to extract the lines for data extraction. Once the lines are obtained, the region of data that gets extracted is specified in the regular expression. The regular expressions were a bit more difficult than I realized but after getting the first two statements down, it became much easier to get it working. There is probably still more room for improvement.