CSV File Processor
All this program really does is take the CSV file created from the bash program and just spits out the data in a more organized manner. The only difference is that I was trying to get used to how linked lists worked, But since this is C, I had to get used to using malloc which was a bit new to me because for the first time, I understood the segmentation fault nightmares were. Also, strtok is pretty neat.